dubai_real_estate_workflow/requirements.txt
# Core AI Framework
crewai==0.28.8
crewai-tools==0.1.6
langchain==0.1.20
langchain-openai==0.1.7
langchain-community==0.0.38

# OpenAI
openai>=1.0.0

# Web Scraping
beautifulsoup4==4.12.3
selenium==4.18.1
scrapy==2.11.1
playwright==1.42.0
requests==2.31.0

# Data Processing
pandas==2.2.1
numpy==1.26.4
openpyxl==3.1.2

# CRM Integrations
gspread==5.12.4
oauth2client==4.1.3
airtable-python-wrapper==0.15.3

# Communication
twilio==9.0.4
httpx==0.27.0

# Validation
phonenumbers==8.13.31
email-validator==2.1.1

# Configuration & Environment
python-dotenv==1.0.1
pyyaml==6.0.1

# Utilities
python-dateutil==2.9.0

# Testing (optional)
pytest==7.4.3
pytest-cov==4.1.0
